% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chyper.R
\name{meanchyper}
\alias{meanchyper}
\title{Mean of conditional hypergeometric distributions}
\usage{
meanchyper(s, n, m)
}
\arguments{
\item{s}{an integer representing the size of the intersecting population}

\item{n}{a vector of integers representing the sizes of each non-intersecting population}

\item{m}{a vector of integers representing the sample sizes}
}
\value{
The mean of the conditional hypergeometric distribution specified
}
\description{
Calculates the mean of a conditional hypergeometric distribution: the distribution of how many items are in the overlap of all samples when samples of arbitrary size are each taken without replacement from populations of arbitrary size.
}
\examples{
meanchyper(10, c(12,13,14), c(7,8,9))

}
