% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chyper.R
\name{mleS}
\alias{mleS}
\title{Maximum likelihood estimator for overlap size in conditional hypergeometric distributions}
\usage{
mleS(k, n, m, verbose = T)
}
\arguments{
\item{k}{the observed overlaps}

\item{n}{a vector of integers representing the sizes of each non-intersecting population}

\item{m}{a vector of integers representing the sample sizes}

\item{verbose}{T/F should intermediate messages be printed?}
}
\value{
The maximum likelihood estimator of the intersecting population size
}
\description{
Calculates the MLE of the overlap size in a conditional hypergeometric distribution: the distribution of how many items are in the overlap of all samples when samples of arbitrary size are each taken without replacement from populations of arbitrary size.
}
\examples{
mleS(c(0,0,1,1,0,2,0), c(12,13,14), c(7,8,9))

}
