% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chyper.R
\name{momM}
\alias{momM}
\title{Method of moments estimator for sample size in conditional hypergeometric distributions}
\usage{
momM(population, k, s, n, m)
}
\arguments{
\item{population}{the index of the unknown sample size}

\item{k}{the observed overlaps}

\item{s}{an integer representing the size of the intersecting population}

\item{n}{a vector of integers representing the sizes of each non-intersecting population}

\item{m}{a vector of integers representing the sample sizes where the value of the unknown sample size should be any integer as a placeholder}
}
\value{
The method of moments estimator of the unknown sample size
}
\description{
Calculates the MOM estimator of a sample size in a conditional hypergeometric distribution: the distribution of how many items are in the overlap of all samples when samples of arbitrary size are each taken without replacement from populations of arbitrary size.
}
\examples{
momM(1, c(0,0,1,1,0,2,0), 8, c(12,13,14), c(0,8,9))

}
