% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chyper.R
\name{rchyper}
\alias{rchyper}
\title{Random number generator for conditional hypergeometric distributions}
\usage{
rchyper(size, s, n, m, verbose = T)
}
\arguments{
\item{size}{the number of random numbers to generate}

\item{s}{an integer representing the size of the intersecting population}

\item{n}{a vector of integers representing the sizes of each non-intersecting population}

\item{m}{a vector of integers representing the sample sizes}

\item{verbose}{T/F should intermediate messages be printed?}
}
\value{
A vector of random numbers generated from the PMF of the conditional hypergeometric distribution specified by the parameters
}
\description{
Generates random numbers from a conditional hypergeometric distribution: the distribution of how many items are in the overlap of all samples when samples of arbitrary size are each taken without replacement from populations of arbitrary size.
}
\examples{
rchyper(100, 10, c(12,13,14), c(7,8,9))

}
